Normed

+++ includes/ft_math.h
@@ -6,7 +6,7 @@
 /*   By: dbaladro <dbaladro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 14:08:17 by dbaladro          #+#    #+#             */
-/*   Updated: 2024/01/20 19:44:32 by madlab           ###   ########.fr       */
+/*   Updated: 2024/01/28 09:40:01 by dbaladro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct s_complex
 	double	img;
 }				t_complex;
 
-unsigned int	ft_abs(int n);
+double			ft_abs(double n);
 t_complex		c_add(t_complex c1, t_complex c2);
 t_complex		c_pow2(t_complex c);
